requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
Keras==2.3.1
Keras-Applications==1.0.8
keras-utils==1.0.13
kiwisolver==1.1.0
numpy==1.16.4
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
pandas==0.25.3
Pillow==7.0.0
requests==2.22.0
requests-oauthlib==1.3.0
scikit-learn==0.22.1
scipy==1.1.0
tensorflow==1.15.2
matplotlib==3.1.2
